requirements.txt
scikit-learn~=0.24.1
numpy~=1.19.5
tqdm~=4.50.2
pandas~=1.2.4
matplotlib~=3.4.1
jupyter
ipykernel
imgaug
opencv-python~=4.5.1.48
torch~=1.8.1+cu111
torchvision~=0.9.1+cu111
Pillow~=8.2.0